v1.0
- SCREEN_CENTER was lifted upwards, and now all functions, that use it
to send troops, work as intended
- Removed templates of major cities, as they have been used only to find
city names, which was purely cosmetic feature and didnt affect the work
of Diana
- Following that, removed get_city_name() function
- For the sake of human reading, all debug screenshots now named in
readable date format, instead of seconds since the start of the era
- Added add_timestamp() function which imprints current time on the
given screenshot
- Interface exclusion, which was scattered throghout the code, now
written in exclude_interface() function
- Added getPos=bool parameter to template_on_image() function, if set to
true, function also returns the center of the given template on image.
Useful for finding UI stuff and then clicking on it
- Added click_if_found() function, which uses that new
template_on_image() feature to click on templates, if they're present on
the image
- As has been proved by debug screenshots, altitude jumps were often
causing decalibration of altitude, so a decision was made to perform all
tasks on altitude level 7. All cv2.HoughCircles() were recalibrated to
work on that level
- Function get_loc_time() now returns YY-MM-DD HH-MM-SS instead of
YYYY/DD/MM HH:MM:SS, because symbols ':' and '/' are not allowed in file
names
- Function get_unique_screenshot_name() renamed to get_screenshot_name()
- Added kill_game(), revive_game(), reboot() functions. Used to restart
the game in case of disconnect or any other critical event. Diana now
can operate 24\7 and restart the game in case of difficulties
- Code that was supposed to close windows appearing after clicking on
towns was rewritten using new click_if_found() function, and now it
actually works